Extract common code to HmdaPersistentActor
