+++ persistence/src/main/scala/hmda/persistence/processing/HmdaRawFile.scala
@@ -1,7 +1,6 @@
 package hmda.persistence.processing
 
-import akka.actor.{ ActorLogging, ActorRef, ActorSystem, Props }
-import akka.persistence.{ PersistentActor, SnapshotOffer }
+import akka.actor.{ ActorRef, ActorSystem, Props }
 import hmda.persistence.CommonMessages._
 import hmda.persistence.LocalEventPublisher
 
@@ -32,7 +31,7 @@ object HmdaRawFile {
 
 }
 
-class HmdaRawFile(submissionId: String) extends PersistentActor with ActorLogging with LocalEventPublisher {
+class HmdaRawFile(submissionId: String) extends HmdaPersistentActor(submissionId) with LocalEventPublisher {
 
   import HmdaRawFile._
 
@@ -42,14 +41,6 @@ class HmdaRawFile(submissionId: String) extends PersistentActor with ActorLoggin
 
   def updateState(event: Event): Unit = {
     state = state.updated(event)
-  }
-
-  override def preStart(): Unit = {
-    log.debug(s"Uploading started for $submissionId")
-  }
-
-  override def postStop(): Unit = {
-    log.debug(s"Uploading finished for $submissionId")
   }
 
   override def receiveCommand: Receive = {
@@ -73,9 +64,4 @@ class HmdaRawFile(submissionId: String) extends PersistentActor with ActorLoggin
       context stop self
   }
 
-  override def receiveRecover: Receive = {
-    case event: Event => updateState(event)
-  }
-
-  override def system: ActorSystem = context.system
 }
